interview-service/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pymongo==4.6.1
groq==0.4.1
openai==1.12.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
reportlab==4.0.9
Pillow==10.2.0
PyMuPDF==1.23.8
# RAG Dependencies
chromadb==0.4.22
tiktoken==0.5.2

